requirements.txt
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==0.8.11
flask==2.3.3
werkzeug==2.3.7
speechrecognition==3.10.0
pyaudio==0.2.11
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1